restructure the packages
