+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/CallScreen.kt
@@ -19,14 +19,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.wire.android.R
 import com.wire.android.ui.common.UserProfileAvatar
-import com.wire.android.ui.main.conversation.all.model.toUserInfoLabel
+import com.wire.android.ui.main.conversation.model.toUserInfoLabel
 import com.wire.android.ui.main.conversation.call.model.Call
 import com.wire.android.ui.main.conversation.call.model.CallEvent
 import com.wire.android.ui.main.conversation.call.model.CallTime
-import com.wire.android.ui.main.conversation.common.MissedCallBadge
-import com.wire.android.ui.main.conversation.common.RowItem
-import com.wire.android.ui.main.conversation.common.UserLabel
-import com.wire.android.ui.main.conversation.common.folderWithElements
+import com.wire.android.ui.main.conversation.common.components.MissedCallBadge
+import com.wire.android.ui.main.conversation.common.components.RowItem
+import com.wire.android.ui.main.conversation.common.components.UserLabel
+import com.wire.android.ui.main.conversation.common.extension.folderWithElements
 import com.wire.android.ui.theme.subLine1
 
 @Composable
