+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/MentionScreen.kt
@@ -14,11 +14,11 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.wire.android.R
 import com.wire.android.ui.common.UserProfileAvatar
-import com.wire.android.ui.main.conversation.all.model.toUserInfoLabel
-import com.wire.android.ui.main.conversation.common.RowItem
-import com.wire.android.ui.main.conversation.common.UnreadMentionBadge
-import com.wire.android.ui.main.conversation.common.UserLabel
-import com.wire.android.ui.main.conversation.common.folderWithElements
+import com.wire.android.ui.main.conversation.model.toUserInfoLabel
+import com.wire.android.ui.main.conversation.common.components.RowItem
+import com.wire.android.ui.main.conversation.common.components.UnreadMentionBadge
+import com.wire.android.ui.main.conversation.common.components.UserLabel
+import com.wire.android.ui.main.conversation.common.extension.folderWithElements
 import com.wire.android.ui.main.conversation.mention.model.Mention
 import com.wire.android.ui.theme.subLine1
