+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/common/components/EventBadge.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.main.conversation.common
+package com.wire.android.ui.main.conversation.common.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
