+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/common/extension/LazyListScope.kt
@@ -1,8 +1,9 @@
-package com.wire.android.ui.main.conversation.common
+package com.wire.android.ui.main.conversation.common.extension
 
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
+import com.wire.android.ui.main.conversation.common.components.FolderHeader
 
 
 inline fun <T> LazyListScope.folderWithElements(
