+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/model/NewActivity.kt
@@ -1,5 +1,7 @@
 package com.wire.android.ui.main.conversation.all.model
 
+import com.wire.android.ui.main.conversation.model.Conversation
+
 data class NewActivity(
     val eventType: EventType,
     val conversation: Conversation,
