+++ app/src/main/kotlin/com/wire/android/ui/main/navigation/MainTopBar.kt
@@ -4,7 +4,6 @@ import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material.DrawerValue
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
@@ -12,10 +11,7 @@ import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.rememberDrawerState
-import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
@@ -23,19 +19,19 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.compose.rememberNavController
 import com.wire.android.R
 import com.wire.android.ui.common.SearchBar
 import com.wire.android.ui.common.UserProfileAvatar
 import com.wire.android.ui.main.WireAppState
+import com.wire.android.ui.main.rememberWireAppState
 import kotlinx.coroutines.launch
 
 @ExperimentalAnimationApi
 @ExperimentalComposeUiApi
 @Composable
 fun MainTopBar(
-    wireAppState: WireAppState,
-    onSearchTextChanged: (String) -> Unit
+    wireAppState: WireAppState
+
 ) {
     with(wireAppState) {
         val currentNavigationScreenItem: MainNavigationScreenItem? = currentNavigationItem
@@ -74,8 +70,9 @@ fun MainTopBar(
                 },
             )
 
+            //TODO:This is a mock search that is going to be further implemented in other sprint's
             if (wireAppState.shouldShowSearchBar) {
-                SearchBar(placeholderText = stringResource(R.string.search_bar_hint, title.lowercase()), onTextTyped = onSearchTextChanged)
+                SearchBar(placeholderText = stringResource(R.string.search_bar_hint, title.lowercase()))
             }
         }
     }
@@ -86,8 +83,5 @@ fun MainTopBar(
 @Preview(showBackground = false)
 @Composable
 fun MainTopBarPreview() {
-    val scope = rememberCoroutineScope()
-    val scaffoldState = rememberScaffoldState(rememberDrawerState(DrawerValue.Closed))
-    val navController = rememberNavController()
-    // MainTopBar(scope = scope, scaffoldState = scaffoldState, navController = navController)
+     MainTopBar(rememberWireAppState())
 }
